[Minor] lua_scanners - move and extend mime_part matching
